gosuc: split incremental compilation into API and implementation
